pyproject.toml
[project]
name = "interview"
version = "0.1.0"
description = "Cheat Sheet"
requires-python = ">=3.8"
dependencies = [
    "requests",
    "python-dotenv",
    "faster-whisper",
    "SpeechRecognition",
    "pyaudio",
    "nvidia-cudnn-cu11",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

